pyproject.toml
[project]
name = "mgapi"
version = "0.1.0"
description = "CLI tool for managing and interacting with MGAPI server"
readme = "README.md"
authors = [
    { name = "kwonah0", email = "kwondkdud@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "inquirerpy>=0.3.0",
    "httpx>=0.24.0",
    "dynaconf>=3.2.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "duckdb>=0.9.0",
    "pandas>=2.0.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
mgapi = "mgapi.cli:main"

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
]
